// MultiModule_Genericservices/component/UserManagement/Models/UserManagement.model.ts
export class UsersManagementModel {
  "staffId":number;
  "floristId": number;
  "userId": number;
  "firstName": string;
  "lastName": string;
  "password": string;
  "mobile": string;
  "phone":string;
  "email": string;
  "dob": string;
  "gridPageSize": string;
  "defultLocation":string;
  "homeScreen":string;
  "createdBy": string;
  "locations": string;
  "artists": any;
  "roles": string;
  "status":string;
  "totalRows": string; 
}
  